% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calldiff_turbo.R
\name{calldiff_turbo}
\alias{calldiff_turbo}
\title{Call Differential Methylation in Windows or Base in Presto}
\usage{
calldiff_turbo(
  met_obj = NULL,
  mean_diff_abs = 0.5,
  groupname = "X",
  diff_group_colname = NULL,
  chunk_size = 50000
)
}
\arguments{
\item{met_obj}{Methylation object containing methylation data and
sample information.}

\item{mean_diff_abs}{Minimum absolute methylation difference
threshold between groups (default: 0.5).}

\item{groupname}{Name for the differential analysis (default: 'X').}

\item{diff_group_colname}{Column name in sample info containing
group labels.}

\item{chunk_size}{Number of rows to process at a time (default: 5e4).}
}
\value{
A data frame containing the results of the differential analysis.
}
\description{
This function calls differential methylation in
windows or base-pairs using the turbo algorithm.
}
\details{
This function uses the turbo algorithm to call differential
methylation in windows or base-pairs. It calculates mean
methylation levels for each group using chunks of data. Then,
it finds significant differences by comparing the absolute
differences between group means for each row. Results are stored
in an HDF5 file with the following structure:
\itemize{
\item \code{uns/diff_results}: Results of the differential analysis.
\item \code{uns/diff_results_idx}: Indices of significant rows.
}
}
\examples{
output_dir <- tempdir()
sam_info_df <- data.frame(
  sample_name = c(
    'sample_1.bed', 'sample_2.bed',
    'sample_3.bed', 'sample_4.bed',
    'sample_5.bed', 'sample_6.bed',
    'sample_7.bed', 'sample_8.bed'
  ),
  group = c(rep('case', 4), rep('ctrl', 4))
)
sam_info <- file.path(output_dir, 'sample_info.csv')

input_file_df <- data.frame(
  marker_name = c(
    'chr1_1000_2000', 'chr1_2000_3000',
    'chr1_3000_4000', 'chr1_4000_5000'
  ),
  sample_1.bed = c(100, 200, 600, 900),
  sample_2.bed = c(100, 200, 600, 900),
  sample_3.bed = c(100, 200, 600, 900),
  sample_4.bed = c(100, 200, 600, 900),
  sample_5.bed = c(800, 900, 100, 50),
  sample_6.bed = c(800, 900, 100, 50),
  sample_7.bed = c(800, 900, 100, 50),
  sample_8.bed = c(800, 900, 100, 50)
)
input_file <- file.path(output_dir, 'methylTracer_1kb.txt')

annotation_file_df <- data.frame(
  chr = rep('chr1', 4),
  start = c(1000, 2000, 3000, 4000),
  end = c(2000, 3000, 4000, 5000),
  SYMBOL = c('gene1', 'gene2', 'gene3', 'gene4'),
  marker_name = c(
    'chr1_1000_2000', 'chr1_2000_3000',
    'chr1_3000_4000', 'chr1_4000_5000'
  )
)
annotation_file <- file.path(output_dir, 'annotation.bed')

output_file <- 'methylTracer_obj_test.h5'
unlink(file.path(output_dir, output_file),
  recursive = TRUE
)
write.csv(sam_info_df, sam_info,
  row.names = FALSE)
write.table(input_file_df, input_file,
  sep = '\t', row.names = FALSE)
write.table(annotation_file_df, annotation_file,
  sep = '\t', row.names = FALSE)

build_h5(
  sam_info = sam_info,
  input_file = input_file,
  output_dir = output_dir,
  output_file = output_file,
  annotation_file = annotation_file
)

met_obj <- build_met_obj(
  file.path(output_dir, output_file),
  sample_name = 'sample_name',
  marker_name = 'marker_name'
)

# Call differential methylation base presto
diff_results <- calldiff_turbo(
  met_obj = met_obj,
  mean_diff_abs = 0.01,
  diff_group_colname = 'group'
)

}
