% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_function.R
\name{computeVar}
\alias{computeVar}
\title{Compute Column Means for Variable Groups}
\usage{
computeVar(met, dataset = "/X", varColName = NULL)
}
\arguments{
\item{met}{A \code{methylTracer} object containing metadata.}

\item{dataset}{A string specifying the dataset path inside the HDF5 file.
It defaults to "/X", but can be customized for other datasets such
as "/coverage".}

\item{varColName}{A string representing the regulation group within
the HDF5 file. This is used to identify groups (or categories) within
the dataset for computing the column-wise means.}
}
\value{
This function does not return any value but writes the computed mean
values for each group into a new dataset in the HDF5 file.
}
\description{
This function computes the mean of selected columns
(variables) from an HDF5 dataset based on specified group indices
in the regulation column (\code{var_group}). It retrieves data from an
HDF5 file, calculates the column-wise means, and writes the result to
a new dataset in the same HDF5 file. The result is saved under a
variable name combining the dataset and regulation group names.
}
\examples{
## input data
suppressMessages(library(HDF5Array))
outputDir <- tempdir()
toy <- system.file('extdata', 'toy.h5', package = 'HDF5Array')
X <- HDF5Array(toy, 'M1')
X <- round(X * 1000, 0)
nRows <- dim(X)[1]
startN <- seq(from = 10000, by = 1000, length.out = nRows)
endN <- startN + 1000

## meta file
marker_name <- paste0('chr1_', startN, '_', endN)
sample_name <- paste0('sample_', 1:dim(X)[2])
group <- c(rep('case', 75), rep('ctrl', 75))
vargroup <- c(rep('intron', 3000), rep('exon', 3000), rep('utr', 4000))
metH5 <- file.path(outputDir, 'methylTracer_obj_test.h5')
unlink(metH5)

## build metH5
rhdf5::h5createFile(metH5)
rhdf5::h5createGroup(metH5, 'obs')
rhdf5::h5createGroup(metH5, 'var')
HDF5Array::writeHDF5Array(X, metH5, 'X')
rhdf5::h5write(marker_name, metH5, 'var/marker_name')
rhdf5::h5write(vargroup, metH5, 'var/vargroup')
rhdf5::h5write(sample_name, metH5, 'obs/sample_name')
rhdf5::h5write(group, metH5, 'obs/group')

## build methylTracer object
met <- build_met_obj(metH5,
    sample_name = 'sample_name',
    marker_name = 'marker_name'
)



## compute var group values
computeVar(met = met, dataset = '/X', varColName = 'vargroup')

}
