% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_scBS_data.R
\name{create_cpg_universe}
\alias{create_cpg_universe}
\title{Create CpG Universe (Genomic Coordinates)}
\usage{
create_cpg_universe(n_sites, verbose = TRUE)
}
\description{
Create CpG Universe (Genomic Coordinates)
}
\keyword{internal}
