% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_utils.R
\name{h5_read_safe}
\alias{h5_read_safe}
\title{Safe HDF5 read operation}
\usage{
h5_read_safe(filepath, dataset_path, ...)
}
\arguments{
\item{filepath}{Path to HDF5 file}

\item{dataset_path}{Path to dataset}

\item{...}{Additional arguments passed to h5read}
}
\value{
Data read from HDF5
}
\description{
Read from HDF5 with guaranteed handle cleanup
}
\keyword{internal}
