% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_utils.R
\name{h5_write_safe}
\alias{h5_write_safe}
\title{Safe HDF5 write operation}
\usage{
h5_write_safe(obj, filepath, dataset_path, ...)
}
\arguments{
\item{obj}{Object to write}

\item{filepath}{Path to HDF5 file}

\item{dataset_path}{Path to dataset}

\item{...}{Additional arguments passed to h5write}
}
\value{
NULL (invisible)
}
\description{
Write to HDF5 with guaranteed handle cleanup
}
\keyword{internal}
