% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_scBS_data.R
\name{implant_dmrs}
\alias{implant_dmrs}
\title{Implant DMRs (Differentially Methylated Regions)}
\usage{
implant_dmrs(
  cpg_universe,
  n_dmrs,
  dmr_width,
  dmr_meth_ctrl,
  dmr_meth_treat,
  verbose = TRUE
)
}
\description{
Implant DMRs (Differentially Methylated Regions)
}
\keyword{internal}
