% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_class.R
\name{met_obs}
\alias{met_obs}
\title{Retrieve Observations from methylTracer Object}
\usage{
met_obs(met_obj = NULL, index = NULL)
}
\arguments{
\item{met_obj}{A \code{methylTracer} object containing methylation
data, which has the file path to the HDF5
file stored in its \verb{@seed@filepath} slot.}

\item{index}{Optional; a numeric vector specifying which rows to retrieve
from the 'obs' dataset in the HDF5 file. If \code{NULL} (default),
the entire 'obs' dataset is retrieved.}
}
\value{
A \code{data.frame} containing the observations from the
'obs' dataset. If \code{index} is specified, only the selected rows
will be returned. If no rows are found or the dataset is empty,
an empty \code{data.frame} is returned.
}
\description{
This function retrieves specified rows or the entire
'obs' dataset from a \code{methylTracer} object stored in an HDF5 file.
The dataset is read using the \code{rhdf5::h5read} function, and the
result is returned as a \code{data.frame}.
}
\examples{
toy_h5 <- system.file('extdata', 'toy.h5', package = 'HDF5Array')
X <- HDF5Array(toy_h5, 'M1') # Load data matrix from HDF5 file

num_rows <- dim(X)[1]
start_values <- seq(from = 10000, by = 1000, length.out = num_rows)
end_values <- start_values + 1000
marker_name <- paste0('chr1_', start_values, '_', end_values)
sample_name <- paste0('sample_', 1:dim(X)[2])

current_dir <- tempdir()
met_obj_test <- paste0(current_dir, '/methylTracer_obj_test.h5')
unlink(met_obj_test) # Remove existing file if it exists

rhdf5::h5createFile(met_obj_test)
rhdf5::h5createGroup(met_obj_test, 'obs')
rhdf5::h5createGroup(met_obj_test, 'var')

HDF5Array::writeHDF5Array(X, met_obj_test, 'X')
rhdf5::h5write(marker_name, met_obj_test, 'var/marker_name')
rhdf5::h5write(sample_name, met_obj_test, 'obs/sample_name')

met_obj_test <- build_met_obj(met_obj_test,
  sample_name = 'sample_name',
  marker_name = 'marker_name'
)

head(met_obs(met_obj_test))
}
