% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_class.R
\name{met_var}
\alias{met_var}
\title{Retrieve Variables from methylTracer Object}
\usage{
met_var(met_obj = NULL)
}
\arguments{
\item{met_obj}{A \code{methylTracer} object containing methylation data,
which has the file path to the HDF5 file.}
}
\value{
A \code{data.frame} containing the first few rows (default: 10)
of each dataset found in the 'var' group of the HDF5 file.
If no data is retrieved, an empty \code{data.frame} is returned.
}
\description{
This function retrieves specified rows (default is the first 10)
from the 'var' datasets within a \code{methylTracer} object stored in
an HDF5 file. It reads datasets under the '/var' group and returns
the extracted rows from each dataset as a \code{data.frame}.
}
\examples{
toy_h5 <- system.file('extdata', 'toy.h5', package = 'HDF5Array')
X <- HDF5Array(toy_h5, 'M1') # Load data matrix from HDF5 file

num_rows <- dim(X)[1]
start_values <- seq(from = 10000, by = 1000, length.out = num_rows)
end_values <- start_values + 1000
marker_name <- paste0('chr1_', start_values, '_', end_values)
sample_name <- paste0('sample_', 1:dim(X)[2])

current_dir <- tempdir()
met_obj_test <- paste0(current_dir, '/methylTracer_obj_test.h5')
unlink(met_obj_test) # Remove existing file if it exists

rhdf5::h5createFile(met_obj_test)
rhdf5::h5createGroup(met_obj_test, 'obs')
rhdf5::h5createGroup(met_obj_test, 'var')

HDF5Array::writeHDF5Array(X, met_obj_test, 'X')
rhdf5::h5write(marker_name, met_obj_test, 'var/marker_name')
rhdf5::h5write(sample_name, met_obj_test, 'obs/sample_name')

met_obj_test <- build_met_obj(
  met_obj_test,
  sample_name = 'sample_name',
  marker_name = 'marker_name'
)

head(met_var(met_obj_test))
}
