% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_scBS_data.R
\name{simulate_single_cell}
\alias{simulate_single_cell}
\title{Simulate a Single Cell's Data}
\usage{
simulate_single_cell(
  cpg_universe,
  dmr_regions,
  group,
  sparsity,
  mean_coverage,
  background_meth
)
}
\description{
Simulate a Single Cell's Data
}
\keyword{internal}
