% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_scBS_data.R
\name{summary.scBS_simulation}
\alias{summary.scBS_simulation}
\title{Summary method for scBS_simulation objects}
\usage{
\method{summary}{scBS_simulation}(object, ...)
}
\description{
Summary method for scBS_simulation objects
}
