% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_count_matrix_SECURE.R
\name{validate_sample_names}
\alias{validate_sample_names}
\title{Validate sample names for safety}
\usage{
validate_sample_names(sams)
}
\description{
Validate sample names for safety
}
\keyword{internal}
