% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_scBS_data.R
\name{write_cell_file}
\alias{write_cell_file}
\title{Write Cell Data to File}
\usage{
write_cell_file(cell_data, cell_id, output_dir, output_format)
}
\description{
Write Cell Data to File
}
\keyword{internal}
