% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_qc.R
\name{impute_met_obj}
\alias{impute_met_obj}
\title{Impute Missing Values in Methylation Data}
\usage{
impute_met_obj(
  met_obj = NULL,
  method = "mean",
  chunk_size = 1e+05,
  level = 1,
  sample_name = NULL,
  marker_name = NULL
)
}
\arguments{
\item{met_obj}{A \code{methylTracer} object containing methylation data.}

\item{method}{The imputation method; options include 'mean' or 'median'.
Default is 'mean'.}

\item{chunk_size}{The size of data chunks processed at a time.
Default is \code{1e5}.}

\item{level}{The compression level for the HDF5 dataset. Default is \code{1}.}

\item{sample_name}{A character string representing the sample names.}

\item{marker_name}{A character string representing the marker names.}
}
\value{
A modified \code{methylTracer} object with imputed values.
}
\description{
This function imputes missing values in a \code{methylTracer} object
using the specified method. The imputed values replace missing data
in the HDF5-stored methylation dataset.
}
\examples{
output_dir <- tempdir()

sam_info_df <- data.frame(
  sample_name = c(
    'sample_1.bed', 'sample_2.bed',
    'sample_3.bed', 'sample_4.bed',
    'sample_5.bed', 'sample_6.bed',
    'sample_7.bed', 'sample_8.bed'
  ),
  group = c(rep('case', 4), rep('ctrl', 4))
)
sam_info <- file.path(output_dir, 'sample_info.csv')

input_file_df <- data.frame(
  marker_name = c(
    'chr1_1000_2000', 'chr1_2000_3000',
    'chr1_3000_4000', 'chr1_4000_5000'
  ),
  sample_1.bed = c(100, 200, 600, 900),
  sample_2.bed = c(100, 200, 600, 900),
  sample_3.bed = c(100, 200, 600, 900),
  sample_4.bed = c(100, 200, 600, 900),
  sample_5.bed = c(800, 900, 100, 50),
  sample_6.bed = c(800, 900, 100, 50),
  sample_7.bed = c(800, 900, 100, 50),
  sample_8.bed = c(800, 900, 100, 50)
)
input_file <- file.path(output_dir, 'methylTracer_1kb.txt')

annotation_file_df <- data.frame(
  chr = rep('chr1', 4),
  start = c(1000, 2000, 3000, 4000),
  end = c(2000, 3000, 4000, 5000),
  SYMBOL = c('gene1', 'gene2', 'gene3', 'gene4'),
  marker_name = c(
    'chr1_1000_2000', 'chr1_2000_3000',
    'chr1_3000_4000', 'chr1_4000_5000'
  )
)
annotation_file <- file.path(output_dir, 'annotation.bed')

output_file <- 'methylTracer_obj_test.h5'

unlink(file.path(output_dir, output_file),
  recursive = TRUE
)

# Write the data to files
write.csv(sam_info_df, sam_info,
  row.names = FALSE
)
write.table(input_file_df, input_file,
  sep = '\t', row.names = FALSE
)
write.table(annotation_file_df, annotation_file,
  sep = '\t', row.names = FALSE
)

build_h5(
  sam_info = sam_info,
  input_file = input_file,
  output_dir = output_dir,
  output_file = output_file,
  annotation_file = annotation_file
)

met_obj <- build_met_obj(
  file.path(output_dir, output_file),
  sample_name = 'sample_name',
  marker_name = 'marker_name'
)

compute_qc_value(
  met_obj = met_obj
)

filter_met_obj <- filter_obs_var(
  met_obj = met_obj,
  obs_cutoff = 0,
  var_cutoff = 0,
  sample_name = 'sample_name',
  marker_name = 'marker_name'
)

met_obj_qc <- impute_met_obj(
  met_obj = filter_met_obj,
  sample_name = 'sample_name',
  marker_name = 'marker_name'
)

}
