% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_met.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot Histogram with Optional Vertical Line}
\usage{
plot_hist(met_obj = NULL, slot = NULL, bins = 50, vline = NULL)
}
\arguments{
\item{met_obj}{A \code{methylTracer} object containing methylation data.}

\item{slot}{A character string specifying the slot in the object
from which the data is read. Example: \code{'obs/coverage_cells'}.
Use \code{methylTracer::compute_qc_values()} to generate slot names.}

\item{bins}{Integer specifying the number of bins for the histogram.
Default is \code{50}.}

\item{vline}{Numeric value specifying the x-position of a vertical line
on the histogram. Default is \code{NULL} (no vertical line).}
}
\value{
A \code{ggplot} object representing the histogram.
}
\description{
This function generates a histogram from a specified slot in
the \code{met_obj} object, with options to customize the number of
bins and add a vertical line at a specified position.
}
\details{
The function reads data from the specified slot in \code{met_obj} and
generates a histogram using \code{ggplot2}. Optionally, a vertical line
can be added at the specified position (\code{vline}).
}
\examples{
output_dir <- tempdir()

sam_info_df <- data.frame(
  sample_name = c(
    'sample_1.bed', 'sample_2.bed',
    'sample_3.bed', 'sample_4.bed',
    'sample_5.bed', 'sample_6.bed',
    'sample_7.bed', 'sample_8.bed'
  ),
  group = c(rep('case', 4), rep('ctrl', 4))
)
sam_info <- file.path(output_dir, 'sample_info.csv')

input_file_df <- data.frame(
  marker_name = c(
    'chr1_1000_2000', 'chr1_2000_3000',
    'chr1_3000_4000', 'chr1_4000_5000'
  ),
  sample_1.bed = c(100, 200, 600, 900),
  sample_2.bed = c(100, 200, 600, 900),
  sample_3.bed = c(100, 200, 600, 900),
  sample_4.bed = c(100, 200, 600, 900),
  sample_5.bed = c(800, 900, 100, 50),
  sample_6.bed = c(800, 900, 100, 50),
  sample_7.bed = c(800, 900, 100, 50),
  sample_8.bed = c(800, 900, 100, 50)
)
input_file <- file.path(output_dir, 'methylTracer_1kb.txt')

annotation_file_df <- data.frame(
  chr = rep('chr1', 4),
  start = c(1000, 2000, 3000, 4000),
  end = c(2000, 3000, 4000, 5000),
  SYMBOL = c('gene1', 'gene2', 'gene3', 'gene4'),
  marker_name = c(
    'chr1_1000_2000', 'chr1_2000_3000',
    'chr1_3000_4000', 'chr1_4000_5000'
  )
)
annotation_file <- file.path(output_dir, 'annotation.bed')

output_file <- 'methylTracer_obj_test.h5'

unlink(file.path(output_dir, output_file),
  recursive = TRUE
)

# Write the data to files
write.csv(sam_info_df, sam_info,
  row.names = FALSE
)
write.table(input_file_df, input_file,
  sep = '\t', row.names = FALSE
)
write.table(annotation_file_df, annotation_file,
  sep = '\t', row.names = FALSE
)

build_h5(
  sam_info = sam_info,
  input_file = input_file,
  output_dir = output_dir,
  output_file = output_file,
  annotation_file = annotation_file
)

met_obj <- build_met_obj(
  file.path(output_dir, output_file),
  sample_name = 'sample_name',
  marker_name = 'marker_name'
)

compute_qc_value(
  met_obj = met_obj
)

plot_hist(
  met_obj = met_obj,
  slot = 'obs/coverage_cells',
  bins = 50,
  vline = 2
)
}
