% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_met.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot Scatter Plot with Grouping Variable}
\usage{
plot_scatter(met_obj = NULL, xcol = NULL, ycol = NULL, group_col = NULL)
}
\arguments{
\item{met_obj}{A methylation object containing the data.}

\item{xcol}{The variable to be plotted on the x-axis (e.g.,
\code{'coverage_cells'}).}

\item{ycol}{The variable to be plotted on the y-axis (e.g.,
\code{'mean_cell_methylation'}).}

\item{group_col}{The variable used for coloring the points (e.g.,
\code{'group'}).}
}
\value{
A \code{ggplot} object representing the scatter plot.
}
\description{
This function generates a scatter plot from a methylation object,
with customizable axes and a grouping variable for coloring points.
}
\examples{
output_dir <- tempdir()

sam_info_df <- data.frame(
  sample_name = c(
    'sample_1.bed', 'sample_2.bed',
    'sample_3.bed', 'sample_4.bed',
    'sample_5.bed', 'sample_6.bed',
    'sample_7.bed', 'sample_8.bed'
  ),
  group = c(rep('case', 4), rep('ctrl', 4))
)
sam_info <- file.path(output_dir, 'sample_info.csv')

input_file_df <- data.frame(
  marker_name = c(
    'chr1_1000_2000', 'chr1_2000_3000',
    'chr1_3000_4000', 'chr1_4000_5000'
  ),
  sample_1.bed = c(100, 200, 600, 900),
  sample_2.bed = c(100, 200, 600, 900),
  sample_3.bed = c(100, 200, 600, 900),
  sample_4.bed = c(100, 200, 600, 900),
  sample_5.bed = c(800, 900, 100, 50),
  sample_6.bed = c(800, 900, 100, 50),
  sample_7.bed = c(800, 900, 100, 50),
  sample_8.bed = c(800, 900, 100, 50)
)
input_file <- file.path(output_dir, 'methylTracer_1kb.txt')

annotation_file_df <- data.frame(
  chr = rep('chr1', 4),
  start = c(1000, 2000, 3000, 4000),
  end = c(2000, 3000, 4000, 5000),
  SYMBOL = c('gene1', 'gene2', 'gene3', 'gene4'),
  marker_name = c(
    'chr1_1000_2000', 'chr1_2000_3000',
    'chr1_3000_4000', 'chr1_4000_5000'
  )
)
annotation_file <- file.path(output_dir, 'annotation.bed')

output_file <- 'methylTracer_obj_test.h5'

unlink(file.path(output_dir, output_file),
  recursive = TRUE
)

# Write the data to files
write.csv(sam_info_df, sam_info,
  row.names = FALSE
)
write.table(input_file_df, input_file,
  sep = '\t', row.names = FALSE
)
write.table(annotation_file_df, annotation_file,
  sep = '\t', row.names = FALSE
)

build_h5(
  sam_info = sam_info,
  input_file = input_file,
  output_dir = output_dir,
  output_file = output_file,
  annotation_file = annotation_file
)

met_obj <- build_met_obj(
  file.path(output_dir, output_file),
  sample_name = 'sample_name',
  marker_name = 'marker_name'
)

compute_qc_value(
  met_obj = met_obj
)

plot_scatter(
  met_obj = met_obj,
  xcol = 'coverage_cells',
  ycol = 'mean_cell_methylation',
  group_col = "group"
)
}
